move user authorization check logic to views.py
Authorized users get access to certain pages of the app that unauthorized users cannot see. So we must check if user is authenticated to grant him access to these pages.

Previously, this check was done in templates. And now the check is moved to views (Class Based Views). By adding LoginRequiredMixin to some CBVs we make sure that only authenticated users will get access to specific pages.

Changes made:

1. views.py
   Add LoginRequiredMixin to PlacesListView, PlaceCreateView, PlaceUpdateView, PlaceDeleteView

2. templates my-places.html, place-form.html, place-update.html, place-delete.html
   "auth-required.html" block removed from templates. The block was responsible for displaying "login" links to the user who is trying to get a page that is accessible only to authorized users

3. account/login.html
   Add login.html template which overrides "allauth" built-in template

4. account/base.html
   Remove all css classes from "body" tag and add some css classes to "main tag".
   These changes make sure that "login" page will be displayed nicely

5. css/base.css
   Add css styles for "Social accounts providers" links (VK, Google) displayed on the "login" page.
   Because these links are generated dynamically by the "accounts/login.html" template they must have the same style as the links on the homepage